tests/wpt/web-platform-tests/payment-request/payment-request-in-iframe.html

<!DOCTYPE html>
<!-- Copyright © 2016 Chromium authors and World Wide Web Consortium, (Massachusetts Institute of Technology, ERCIM, Keio University, Beihang). -->
<meta charset="utf-8">
<title>Test for PaymentRequest in an iframe (see also
https://github.com/w3c/browser-payment-api/issues/2)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe srcdoc="
<!DOCTYPE html>
<meta charset='utf-8'>
<script>
window.top.test(function() {
    window.top.assert_throws({name: 'SecurityError'}, function() {
        new PaymentRequest([{supportedMethods: ['foo']}], {total: {label: 'label', amount: {currency: 'USD', value: '5.00'}}});
    }, 'If the browsing context of the script calling the constructor is not a top-level browsing context, then throw a SecurityError.');
});
</script>
"></iframe>
